Fix dependency collection for mixed workspaces
`cargo geiger` has not worked well in workspaces that have a top-level
package and at least one member package. It would invoke a `cargo check`
only on the top-level package, but attempt a source analysis on the
dependencies of all packages the workspace contained, which would fail
due to the search step not being able to retrieve sources of the
dependencies that have not been `cargo check`ed.

This change fixes the problem, however, leaves open the question of how
multi-package workspaces (especially the ones with a top-level package,
i.e. with a non-virtual manifest) should be handled by the tool.